requirements.txt
ultralytics>=8.0.0
rich>=13.0.0
bokeh>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
click>=8.1.0
plotly>=5.15.0
psutil>=5.9.0
tensorboard>=2.13.0
wandb>=0.15.0
